adding  sdrInput and inputBase classes
