+++ boswatch/inputSource/__init__.py
@@ -1,0 +1,2 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
